エラー修正完了

+++ src/servlet/ParentsServlet.java
@@ -84,11 +84,10 @@ public class ParentsServlet extends HttpServlet {
 			String userID = loginUser.getUserId();
 			String date1 = request.getParameter("date");
 			java.sql.Date date= java.sql.Date.valueOf(date1);
-			String comment = new String(request.getParameter("comment").getBytes("ISO-8859-1"),"UTF-8");
+			String comment =request.getParameter("comment");
 			CommentDao coDao = new CommentDao();
 			// submit パラメータの文字コードの問題を修正するため、値を直接比較する
-			String submitValue = request.getParameter("submit");
-			if ("登録".equals(submitValue)) {
+			if (request.getParameter("submit").equals("登録")) {
 	            if (coDao.insert(new CalendarComment(date, userID, comment))) {
 	                request.setAttribute("result",
 	                        new Result("登録成功！", "本日のコメントを登録しました😊", "/D2/ParentsServlet"));
@@ -97,15 +96,6 @@ public class ParentsServlet extends HttpServlet {
 	                        new Result("登録失敗…", "コメントを登録できませんでした😢", "/D2/ParentsServlet"));
 	            }
 	        }
-			if (new String(request.getParameter("submit").getBytes("ISO-8859-1"),"UTF-8").equals("登録")) {
-			    if (coDao.insert(new CalendarComment(date, userID, comment))) {
-			        request.setAttribute("result",
-			                new Result("登録成功！", "本日のコメントを登録しました😊", "/D2/ParentsServlet"));
-			    } else {
-			        request.setAttribute("result",
-			                new Result("登録失敗…", "コメントを登録できませんでした😢", "/D2/ParentsServlet"));
-			    }
-			}
 		} else if (action.equals("housework_regist")){
 			HttpSession session = request.getSession();
 			User loginUser = (User) session.getAttribute("id");
@@ -116,12 +106,12 @@ public class ParentsServlet extends HttpServlet {
 
 			request.setCharacterEncoding("UTF-8");
 
-			String houseworkName = new String(request.getParameter("houseworkName").getBytes("ISO-8859-1"),"UTF-8");
-			String houseworkContents = new String(request.getParameter("houseworkContents").getBytes("ISO-8859-1"),"UTF-8");
+			String houseworkName = request.getParameter("houseworkName");
+			String houseworkContents =request.getParameter("houseworkContents");
 			String houseworkPoint = request.getParameter("houseworkPoint");
 
 			HouseworkDao wDao = new HouseworkDao();
-			if (new String(request.getParameter("submit").getBytes("ISO-8859-1"),"UTF-8").equals("更新")) {
+			if (request.getParameter("submit").equals("更新")) {
 				if (wDao.updateHW(new HouseWork(houseworkName, houseworkContents, houseworkPoint,userID))) {
 					request.setAttribute("result",
 							new Result("更新成功！", "更新を実施しました", "/D2/ParentsServlet"));
@@ -129,7 +119,7 @@ public class ParentsServlet extends HttpServlet {
 					request.setAttribute("result",
 							new Result("更新失敗…","更新出来ませんでした", "/D2/ParentsServlet"));
 				}
-			}else {
+			}else if(request.getParameter("submit").equals("削除")){
 				if(wDao.delete(houseworkName)) { // 削除成功
 					request.setAttribute("result",
 							new Result("削除成功！","削除を実施しました", "/D2/ParentsServlet"));
@@ -146,22 +136,21 @@ public class ParentsServlet extends HttpServlet {
 
 
 			request.setCharacterEncoding("UTF-8");
-			int childId = Integer.parseInt(request.getParameter("childId"));
 			Part childPicture = request.getPart("childPicture");
 			if (childPicture.getSubmittedFileName() != "") {
 				childPicture.write("C:\\pleiades\\workspace\\D2\\WebContent\\upload\\"+ childPicture.getSubmittedFileName());
 			} 
-			String childName = new String( request.getParameter("childName").getBytes("ISO-8859-1"),"UTF-8");
+			String childName =request.getParameter("childName");
 			//String userId = request.getParameter("userId");
 			String rewardUmu = request.getParameter("rewardUmu");
 			String rewardJouken = request.getParameter("rewardJouken");
-			String rewardText = new String( request.getParameter("rewardText").getBytes("ISO-8859-1"),"UTF-8");
+			String rewardText = request.getParameter("rewardText");
 
 
 			ChildDao dDao = new ChildDao();
 
-			 if(new String(request.getParameter("submit").getBytes("ISO-8859-1"),"UTF-8").equals("登録")){
-				if (dDao.insert(new Child(childId, childPicture.getSubmittedFileName(), childName, userID, rewardUmu, rewardJouken, rewardText))) { // 登録成功
+			 if(request.getParameter("submit").equals("登録")){
+				if (dDao.insert(new Child(0, childPicture.getSubmittedFileName(), childName, userID, rewardUmu, rewardJouken, rewardText))) { // 登録成功
 					request.setAttribute("result",
 							new Result("登録成功！", "登録を実施しました", "/D2/ParentsServlet"));
 				} else { // 更新失敗
@@ -173,30 +162,30 @@ public class ParentsServlet extends HttpServlet {
 			HttpSession session = request.getSession();
 			User loginUser = (User) session.getAttribute("id");
 			String userID = loginUser.getUserId();
-
+			int childId = Integer.parseInt(request.getParameter("childId"));
 			request.setCharacterEncoding("UTF-8");
 			Part childPicture = request.getPart("childPicture");
 			if (childPicture.getSubmittedFileName() != "") {
 				childPicture.write("C:\\pleiades\\workspace\\D2\\WebContent\\upload\\"+ childPicture.getSubmittedFileName());
 			} 
-			String childName = new String( request.getParameter("childName").getBytes("ISO-8859-1"),"UTF-8");
+			String childName = request.getParameter("childName");
 			//String userId = request.getParameter("userId");
 			String rewardUmu = request.getParameter("rewardUmu");
 			String rewardJouken = request.getParameter("rewardJouken");
-			String rewardText = new String( request.getParameter("rewardText").getBytes("ISO-8859-1"),"UTF-8");
+			String rewardText = request.getParameter("rewardText");
 
 
 			ChildDao dDao = new ChildDao();
 
-			if (new String(request.getParameter("submit").getBytes("ISO-8859-1"),"UTF-8").equals("更新")) {
-				if (dDao.update(new Child(0, childPicture.getSubmittedFileName(), childName, userID, rewardUmu, rewardJouken, rewardText))) { // 更新成功
+			if (request.getParameter("submit").equals("更新")) {
+				if (dDao.update(new Child(childId, childPicture.getSubmittedFileName(), childName, userID, rewardUmu, rewardJouken, rewardText))) { // 更新成功
 					request.setAttribute("result",
 							new Result("更新成功！", "更新を実施しました", "/D2/ParentsServlet"));
 				} else { // 更新失敗
 					request.setAttribute("result",
 							new Result("更新失敗…","更新出来ませんでした", "/D2/ParentsServlet"));
 				}
-			} else {
+			} else if(request.getParameter("submit").equals("削除")){
 				if (dDao.delete(childName)) { // 削除成功
 					request.setAttribute("result",
 							new Result("削除成功！","削除を実施しました", "/D2/ParentsServlet"));
